Visual updates
For games that haven't started, no longer show the score. Start time is wide enough to keep all logos aligned.
Quit button is no longer an X in the top right. Moved to a gear icon in the bottom right that opens a dropdown (can add other settings there later).