Added NFL
Settings now includes "Change Sport" menu that lets the user flip between MLB and NFL.
When NFL is selected, menu bar icon changes to a football.

For NFL, I will need to adjust display a bit. Not all games are same day, so should include date and time for scheduled game. Also, just the quarter isn't enough for football, we want time remaining in the quarter (will be useful when we add NHL/NBA).